gst-plugins0.10: Remove support for building single plugins from base/good/bad/ugly
Support for building a single plugin from the base/good/bad/ugly
tarballs was originally done to mimic Gentoo's packages.  This makes
sense for a source-based distro, but for a binary package distro it is
horribly inefficient wrt disk space (requiring a -src package for every
plugin).  Instead, just build as much as possible and create separate
binary packages for each plugin with external deps.

This removes the need for GST_PLUGINS_SRC and GST_PLUGINS_DIR, which
were not previously documented with this deprecation in mind.