__list_deps: Search only for DLLs in mono dllmaps
Mono dllmaps can be specified per platform, e.g.:

  <dllmap dll="foo.dll" target="libfoo.so.2" os="!osx,windows,cygwin" />
  <dllmap dll="foo.dll" target="libfoo.2.dylib" os="osx" />
  <dllmap dll="foo.dll" target="cygfoo-2.dll" os="cygwin" />

In this case, we only want to search for the Cygwin DLL to determine
dependencies.  Since os=windows doesn't use dllmaps (the Windows DLL
name is supposed to be the one used in code, for compatibility with .NET
which does not support dllmap-ing), making sure that target= ends in .dll
should suffice.