cross: complete overhaul
Building cross-compilers is now the job of toolchain.cygclass, and the
fragile automatic host/target detection was removed.  The new system is
based on GCC sysroots and will rely on hooks throughout the code to
accomodate the differences between native- and cross-compiling.

docrossbin and docrosslib were removed; dobin and dolib provide the same
functionality when using cross.cygclass.  A doelflib function was added
for installing ELF libraries and their symlinks.