cyginstall: install into CROSS_SYSROOT when cross-compiling
Cross-compiled packages are configured with their usual prefix (/usr on
most systems) so they are built completely ignorant of the sysroot.
Therefore they must be forced into the sysroot during install.

Note that there are some known bugs in libtool wrt sysroots, which can
cause relinks (on ELF platforms) to fail due to spurious -L/usr/lib
flags, which result in a failed attempt to link against Cygwin-native
libraries.  These will need to be addressed soon, otherwise we will have
to rethink this approach to sysroots.