cygautoreconf: set aclocal.m4 to be ignored if nonexistant
A package may use aclocal even if aclocal.m4 is not present, for example
when buiding a VCS snapshot.  In that case, it is safe to ignore
aclocal.m4 in pkg_diff: either it will never exist, or it will be
created during autoreconf and should rightly be ignored.

The only catch would be a package which uses AC_CONFIG_SUBDIRS where one
level uses an aclocal-generated aclocal.m4 and another uses a handwritten
aclocal.m4.  Whether such a case exists in practice will have to be seen.