Enforce standard compiler settings
This removes the ability to modify *FLAGS in cygport.conf.  While it was
useful before gcc4 went stable, it was meant to allow individual users
to rebuild their packages, and it is doubtful that this ever happens.
In order to have uniformity in the distro, we enforce the defaults
but leave individual packages the ability to change these as necessary.