gtk2-perl: delay manpage renaming to postinstall
Apparently the 'manifypods' target has been added to 'all' in EU:MM,
which is a dependency of the 'install' target.  Renaming the manpages
during the build results in 'manifypods' re-running during 'install',
resulting in two sets of manpages (one with '.' and another with '::').

Instead, we now rely on __prepman() to handle manpage renaming (from the
previous commit).  This means gtk2_perl_compile() does nothing more than
perl_compile(); we leave the former in place for API compatibility, but
we no longer need another src_compile() override.