cmake: silence "CMake no longer defines WIN32 on Cygwin!" warning
CMake 2.8.4 dropped the WIN32 define on Cygwin (Ports' cmake package
had removed it ages ago), but emits a warning instead that this change
may affect or break the building of the package in question.

As we've been building packages w/o WIN32 for all this time, this
warning is irrelevant, so just silence it.