Do not use pbzip2 for compressing Cygwin packages
Commit 6f1f87eeb9201990138c9db16f2653c6944926c2 added support for using
pbzip2 for BZ2 compression.  Unfortunately setup.exe does not properly
support these yet, so we need to use only bzip2 in this case.

However, the entire commit need not be reverted, as it also affected
VCS checkout tarball compression, the decompression of which is within
the purview of cygport, not setup.exe.