autotools: forcefully disable /usr/bin/foo-config detection when cross-compiling
If a foo-config script is missing from CROSS_BINDIR but is present in
/usr/bin, the latter will be found even though it is for the wrong
toolchain; use of this result will eventually lead to a compilation or
linking error.

Therefore it is not enough to specify foo-config locations which are in
CROSS_BINDIR, but disable detection of those only in /usr/bin as well.

The FOO_CONFIG variables are also exported in config.cache form
(ac_cv_path_FOO_CONFIG) because many times the given variable is not
marked as special by AC_ARG_VAR.