gst-plugins: Unite and add support for the 0.11/1.0 series
As plugins are often moved between packages during a series, it is best
left to .cygport files to determine which plugins to enable or disable
for a given version.  With that, together with no longer trying to easily
support building a single plugin from the major plugin packages, a single
cygclass can support multiple branches.

gst-plugins0.10.cygclass and GST_PLUGINS_EXT remain only for backwards
compatibility.  The only catch is that the latter will no longer
automatically disable unspecified plugins in the package.