pkg_dist: add support for setup.hint autogeneration
If [PKG_]CATEGORY, [PKG_]SUMMARY, and [PKG_]DESCRIPTION are defined,
then we have enough information to autogenerate a setup.hint file if one
was not provided. We rely on __list_deps to fill in the requires: field,
so [PKG_]REQUIRES is optional.

This is limited to a native build, as this uses cygcheck and the build
system's package database.