Stop trying to canonicalize triplets
There are several customary abbreviations to system triplets:
1) -unknown-none (bare metal) is usually dropped (e.g. avr, msp430)
2) -unknown- is sometimes dropped on embedded systems (e.g. arm-eabi),
but not for Linux (e.g. x86_64-unknown-linux-gnu, armv7hl-unknown-linux-gnu),
as by doing so would be misunderstood as a GNU Hurd system.
3) -none may also be dropped.

The problem with that is on 64-bit Cygwin, which config.guess defaults
to x86_64-unknown-cygwin (even though we prefer x86_64-pc-cygwin), but
x86_64-cygwin is a legal abbreviation.  So even though it technically
breaks API, stop trying to canonicalize names and require consistent use
of TOOLCHAIN_TARGET and CROSS_HOST.