prepstrip: create split debuginfo only if symbols exist therein
Even if a binary is compiled without -g, some debug symbols will exist
from the static parts of libcygwin.a or libgcc.a, or any static libraries
against which it was linked.  So objcopy(1) will find something to do,
but without corresponding source files the split debuginfo is worthless.
If the entire package is so compiled, cygport will not create a -debuginfo
subpackages, and those files will show up as unpackaged.