prepstrip: check for 64bit DLLs with 32bit ImageBase
As --enable-auto-image-base became the default only in the last few years,
some packages use a -Wl,--image-base or BASE .def file directive to set
an ImageBase.  We don't want this on x86_64, though, as those DLLs are
slotted to go in the upper memory area.

This check adds a dependency on gawk.