perl: move cleanup functions to automatic postinstall
Both perllocal.pod and .packlist files are used only within the context
of CPAN-managed modules.  We were already deleting the former due to
clobbering issues (it is supposed to be a cumulative list) and fixing
up the latter to remove the $D prefix.

The problem now arises with multiple arches, in that .packlist is created
under PERL_VENDORARCH, causing even pure Perl (noarch) modules to use
an arch-specific directory, where the same package should really be usable
on both arches.

Therefore, we now follow Fedora's lead and remove .packlist files, as
well as moving all of this to be automatic, as we don't need anything
from perl.cygclass to make this work (as long as the Perl installation
setup does not dramatically change in the future).