autotools: fix for configure-generated-file-in-srcdir check
When building in a directory other than $B, such as in a multilib scenario,
the "Removing $f to be regenerated by configure" operation must use PWD
instead of $B to point to where the file will actually be created.  That
does mean that the symlink may be changed multiple times per compile,
but if that breaks something, further intervention will be required anyway.