Remove support for KDE3
KDE3 is long gone, and cannot be installed in the same prefix as KDE4.
While TDE may be a replacement, currently (3.5.13.x) it too collides
with KDE4, and is in the process of switching to CMake.  Once their
renaming process is complete (due in the next major release), adding
TDE can be reconsidered, but the changes necessary to support it will
necessitate a new cygclass anyway.