prepstrip: use file(1) to identify .Net assemblies
'objdump -p' can take too long on large (usually C++)_unstripped binaries.
file(1) checks certain bytes of the PE header which distinguish .Net
assemblies from regular PE binaries, which is very fast.  The only possible
disadvantage is that the text description may change in the future; if that
becomes a concern, the file(1) check should be reimplemented in cygport via
hexedit or xxd.