pkg_pkg: use skip tag for source-only setup.hint
One noticable difference between upset and genini is the handling of
source-only packages.  genini causes them not to be listed in setup.ini,
but external-source: still works as intended.  upset has included them
in the list, confusing users who think they're installing a library, and
instead getting the sources (if they even notice that).  I prefer genini's
behaviour (often publicly) in this case.

The compromise was a recent addition[1] of the skip: tag, which will
cause upset not to list that package.  In the case that we are auto-
generating setup.hint files (IOW the necessary variables are defined,
and no such file already exists in $C), then use it.

(genini will need to be taught about this tag ASAP.)

[1] http://cygwin.com/ml/cygwin-apps/2013-09/msg00012.html