autotools: assure automake-1.9 with libtool
libtool-2.x requires at least automake 1.9, so if the package ships with
an older version and WANT_AUTOMAKE is unset or set too low, override in
order to assure that autoreconf will succeed.  Otherwise, one has to
explicitly set WANT_AUTOMAKE in each such cygport(3), and then make sure
to remove that when a later version of the package ships with a newer
version of automake.