Do not replace symlinks with copies for MinGW/MSYS/DJGPP
Many symlinks that would end up in a sysroot are primarily intended for
cross-compiling (e.g. headers, implibs, and .pc files; notable exceptions
are icon themes and images in GNOME/MATE help docs), which is the primary
purpose of sysrooted packages, not running the resulting executables, so
we shouldn't need to replace all symlinks with copies even for PE targets.