Add -Wimplicit-function-declaration to CFLAGS/OBJCFLAGS
A common cause of SEGVs on x86_64-cygwin is unprototyped function calls,
as the C standard dictates that the return type be inferred to be 'int'.
While this isn't much of a practical problem on x86 (since sizeof(int) ==
sizeof(long) == sizeof(void *)), it very much is on x86_64 when the return
type is really a 'long' or a pointer (which are larger than 'int').

Adding -Wimplicit-function-declaration to the default CFLAGS should make
it easier to detect potential crashers.  This flag is only for C/ObjC;
with C++ and Java, this is already an automatic error.