postinst: create relevant postinstall commands for cross-compile testing
As we can actually use the mingw* sysroots at runtime for testing
purposes, it makes sense to install runtime data and make it usable.

This will create the necessary postinstall commands for the following
types of data: desktop menu entries, mimetype entries, GConf and GSettings
schemas, and icon themes.  These commands all use the standard system tools.
GDK-Pixbuf and Pango loaders are not handled (yet), as they require the
cross-compiled tools to install; for now, create postinstalls for such
packages, but be sure to account for x86_64-mingw targets on i686-cygwin.

For other cross-hosts, none of this is useful, so just don't include
those files in the package.