prep_texlive: group enable/disable commands in calls to updmap
In postinstall scripts we remove the updmap calls completely and
instead create a file of commands in a marker directory, to be run by
a perpetual postinstall script.  In preremove scripts we just group
the commands into a single call to updmap, since setup.exe doesn't yet
support perpetual preremove scripts.