prep_xsessions: set desktop variables usually handled by DM
DEs expect certain variables to be set -- DESKTOP_SESSION,
XDG_SESSION_DESKTOP, and/or the new XDG_CURRENT_DESKTOP -- and they seem
to rely on the DM to set these variables rather than their SM.  In fact,
GNOME Flashback doesn't even start properly unless the latter is set.
Therefore, we need to set these variables before starting the SM.

As that makes a command line which seems to be too long for a Windows
shortcut, so we create Xclients.$NAME.sh scripts instead, and modify
the shortcuts accordingly.