list_deps: fix shebang detection when multiple scripts are present
This fixes a regression in commit 3860996b2bab250d8ac86672485b03ca79359a01
that caused packages with e.g. a postinstall script (which has no shebang)
and a script with a shebang from finding the dependency of the latter.
With a number of lines to scan specified but no -s, only the first line of
the first script would be scanned.