Propagate exit status
The commands compile, check, inst, package and all currently return the exit
status of the last command in the pipeline, tee, which they pass their
output through.

I guess we don't want to just turn on pipefail, since pkg_pkg very carefully
turns it on and off in a particular place, so use $PIPESTATUS to propagate
the much more interesting exit status of the command who's output is tee'd.

Signed-off-by: Jon Turney <jon.turney@dronecode.org.uk>