Pass BUILD_DEPENDS through to .hint for source package as build-depends:
Converting a dependency atom to a package name with full generality requires
a database of all the pathnames contained in all packages, so we don't even
try to do that here.

In the fullness of time, calm may have the information needed to perform
that task, at which time it can start emitting build-depends: lines for
source packages into setup.ini.