Canonicalize whitespace appropriately in .hint file values
People like to insert extra whitespace (including newlines) into the
variables which control these, for clarity and formatting in the cygport
file. This is a resonable thing to do.

Canonicalize any whitespace in the values of .hint file keys which are
single-line, space-separated lists, to avoid those newlines leaking
through into the .hint file, making the value multi-line, and hence
invalid.

Update tests where the expected hints currently contain excess
whitespace (usually at the end of the line) appropriately.