python-wheel: Handle projects without a setup.py
Since setuptools 40.9.0, projects can follow PEP 517, and have just a
setup.cfg, and a default setup.py is emulated.

In that case, rather than setuptools, use 'pip wheel' to build the
wheel.

Note: possibly could use the pypa build module to do the build, but we
need pip for the install anyhow?