Ignore errors in nested source-ing of cygport
This avoids immediately terminating on calling the 'exit' built-in after
any error in the source-ing (e.g. raised by free code in an inherited
cygclass), which causes a silent exit now we that suppress stderr from
that nested source-ing.

We do this by writing "( stuff-which-may-exit ) || true".

You might think that the extra subshell is not needed.  I certainly did,
(because the rules for the 'errexit' mode (which the subshell inherits)
are that non-final commands in || lists don't cause an exit if they
fail).  I was baffled for a long time as to why it's needed until I
realized that the 'exit' built-in causes the shell to exit, even when
it succeeds.

The '|| true' ensures that the substitution does not have a non-zero
exit status, which is an error in the top-level shell.

Fixes: aa665473aa7d ("Send informational messages to stderr")