Add mechanism to allow cygclasses to register hooks for each phase
Checks in free code in cygclasses are ill-advised: They get run whatever
the subcommand is, and for nested source-ing of the cygport.

Add a mechanism to allow cygclasses to register hooks, which can run on
a particular subcommand.  Use that to defer some compilation environment
checks in cross and XXX cygclasses until the compilation phase.

Also allow a user hook to be registered for that mechanism.