Update tests for python2 EOL
* autotools/mate-menus: Disable the python2 bindings

The bindings in this version aren't compatible with python3.  Later
versions, which are python3 compatible, use introspection-based, rather
than static bindings (so there's nothing to package), so just disable
building the python2 bindings, for the moment.

* download/svn: Drop making a python 2.7 wheel

* python/distutils: BeautifulSoup 3.x doesn't support python3, so
upgrade to bs4

(not sure what to do when distutils goes away, and this becomes another
test of wheel, but still needs a test for the moment)

* python/pygtk: Drop test

Drop this, snce pygtk is python2 (and gtk+2.0) only, there's nothing
left to test.

* python/wheel: Drop making a python 2.7 wheel

* waf/all: Update version of lv2

Update to version of lv2 which still uses waf, but a waf which is
compatible with python3

* Don't install python2 test prerequisites in CI.