src_postinst: Improve X session shortcut folder naming
Don't create shortcuts (but still remove them) if CYGWIN_SETUP_OPTIONS
contains "no-startmenu".

If CYGWIN_START_MENU_SUFFIX is defined, it overrides the old logic for
determining a suffix (which is " (32-bit)" for WOW64) (but still keep
that logic for compatibility with old setup versions).

The idea here is 1) x86_64 setup running on ARM64 will set
CYGWIN_START_MENU_SUFFIX appropriately, and 2) we might even allow the
suffix to be set via a setup option, so that multiple installations
dont't fight over these shortcuts.

Future work: generic helper for a cygport to create these shortcuts?