Improve error if CATEGORY or SUMMARY isn't set
At the moment, if CATEGORY or SUMMARY aren't set, it's assumed that this
is an old .cygport with explicit, handwritten .hint files, which we warn
about being missing.

Update the message, since the contemporary assumption is that .hint
files will be auto-generated, and make it fatal, since the package is
never going to be useful without a hint.

Also: Fix a typo in link to category list

Future work: Stop supporting handwritten .hints completely?