lib/src_postinst.cygpart: parallelize __prepstrip
Factor out stripping one file as __prepstrip_one. The continues
statements from the loop body need to be changed to return 0 statements.

Exchange the pipe into read construct with a process substitution, so a
subshell isn't used for the pipeline and we can wait for the
__prepstrip_one subprocesses to complete.

Prevent output races by splitting the output from each background
process into separate files, avoid filename collisions by using the
ordinal for the in-process file.  Instead of sorting from a single file,
sort from the (many) split files when collecting the debug sources.

Send debug output for prepstip to a separate file.