lib/src_prep.cygpart: Fix incorrect options passed to unzstd in unpack()
Fix incorrect options passed to unzstd in unpack() in lib/src_prep.cygpart.

`unpack()` invokes `unzstd` for a *.zst file as `unzstd -qo file.zst`.
Then, an error occurs as 'stdin is a console, aborting', because the
`-o` option is to specify an output file name, which is missing.

Also handle the case of .xz compressed files.