python-wheel: Check for pythonXX-wheel
The python wheel package is required for building using the python-wheel
cygclass, but nothing in cygport verifies its existence, and the error
from the Python commands themselves aren't particularly helpful either.
To avoid other people wasting the time I just did trying to debug Python
build errors, check if the relevant wheel executable is installed, and
complain if it isn't.

(This seems like it might be an oversight in 2596f685, since pip depends
on setuptools but not wheel)