Add repro-build, repro-diff und repro-check commands
These commands check for reproducibility of distribution packages.

The repro-build command unpacks the source package from the dist
directory to the temp directory and performs a nested rebuild of the
packages there.

The repro-diff command compares original and rebuild packages. If
different, a report about individual differing files in dist, inst and
spkg directories is printed.

The repro-check command combines both commands.