meson: Use '_build.${CHOST}' rather than '${CHOST}' as builddir
Use '_build.${CHOST}' rather than '${CHOST}' as the name of the build
directory.

It seems that a cross-pkgconf will try to use a directory named after
the target triple as the sysroot, if it finds one in the working
directory, which we definitely don't want!

Based on a patch by Carlo Bramini.

https://cygwin.com/pipermail/cygwin-apps/2025-March/044077.html